+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.9.5](https://github.com/AlexsJones/llmfit/compare/v0.9.4...v0.9.5) (2026-04-11)
+
+
+### Features
+
+* **plan:** KV cache fidelity + KV quant flag + TurboQuant gating ([1f1b6fc](https://github.com/AlexsJones/llmfit/commit/1f1b6fc25a04ee1cd7881f4e33becf7d24b95588))
+
+## [0.9.4](https://github.com/AlexsJones/llmfit/compare/v0.9.3...v0.9.4) (2026-04-11)
+
+
+### Bug Fixes
+
+* **download:** fetch all shards of multi-part GGUF models ([b7494d9](https://github.com/AlexsJones/llmfit/commit/b7494d9910dff9e515337a6af9fc2acad59a2628))
+
 ## [0.9.3](https://github.com/AlexsJones/llmfit/compare/v0.9.2...v0.9.3) (2026-04-09)